requirements.txt
PySide6
QT-PyQt-PySide-Custom-Widgets
pyinstaller
reportlab
matplotlib
pandas
openpyxl
PyMuPDF
Pillow
pytesseract
libratom
mammoth
xlrd
odf